+++ projects/js-packages/idc/components/idc-screen/index.jsx
@@ -14,6 +14,7 @@ import trackAndBumpMCStats from '../../tools/tracking';
 import useMigration from '../../hooks/use-migration';
 import useMigrationFinished from '../../hooks/use-migration-finished';
 import useStartFresh from '../../hooks/use-start-fresh';
+import customContentShape from '../../tools/custom-content-shape';
 
 /**
  * The IDC screen component.
@@ -24,7 +25,7 @@ import useStartFresh from '../../hooks/use-start-fresh';
 const IDCScreen = props => {
 	const {
 		logo,
-		headerText,
+		customContent,
 		wpcomHomeUrl,
 		currentUrl,
 		apiNonce,
@@ -32,6 +33,7 @@ const IDCScreen = props => {
 		redirectUri,
 		tracksUserData,
 		tracksEventData,
+		isAdmin,
 	} = props;
 
 	const [ isMigrated, setIsMigrated ] = useState( false );
@@ -76,7 +78,7 @@ const IDCScreen = props => {
 	return (
 		<IDCScreenVisual
 			logo={ logo }
-			headerText={ headerText }
+			customContent={ customContent }
 			wpcomHomeUrl={ wpcomHomeUrl }
 			currentUrl={ currentUrl }
 			redirectUri={ redirectUri }
@@ -87,6 +89,7 @@ const IDCScreen = props => {
 			isFinishingMigration={ isFinishingMigration }
 			isStartingFresh={ isStartingFresh }
 			startFreshCallback={ startFreshCallback }
+			isAdmin={ isAdmin }
 		/>
 	);
 };
@@ -94,8 +97,8 @@ const IDCScreen = props => {
 IDCScreen.propTypes = {
 	/** The screen logo. */
 	logo: PropTypes.object,
-	/** The header text. */
-	headerText: PropTypes.string,
+	/** Custom text content. */
+	customContent: PropTypes.shape( customContentShape ),
 	/** The original site URL. */
 	wpcomHomeUrl: PropTypes.string.isRequired,
 	/** The current site URL. */
@@ -110,6 +113,12 @@ IDCScreen.propTypes = {
 	tracksUserData: PropTypes.object,
 	/** WordPress.com event tracking information. */
 	tracksEventData: PropTypes.object,
+	/** Whether to display the "admin" or "non-admin" screen. */
+	isAdmin: PropTypes.bool.isRequired,
+};
+
+IDCScreen.defaultProps = {
+	customContent: {},
 };
 
 export default IDCScreen;