+++ projects/packages/search/tests/php/test-module-control.php
@@ -111,7 +111,12 @@ class Test_Module_Control extends TestCase {
 	public function test_enable_instant_search() {
 		delete_option( Module_Control::SEARCH_MODULE_INSTANT_SEARCH_OPTION_KEY );
 		static::$search_module->enable_instant_search();
+		// plan doesn't support instant search.
+		$this->assertFalse( get_option( Module_Control::SEARCH_MODULE_INSTANT_SEARCH_OPTION_KEY ) );
+		add_filter( 'jetpack_options', array( $this, 'return_search_active_array' ) );
+		static::$search_module->enable_instant_search();
 		$this->assertTrue( get_option( Module_Control::SEARCH_MODULE_INSTANT_SEARCH_OPTION_KEY ) );
+		remove_filter( 'jetpack_options', array( $this, 'return_search_active_array' ) );
 	}
 
 	/**