+++ projects/plugins/backup/src/js/components/Admin.js
@@ -6,8 +6,11 @@ import { __ } from '@wordpress/i18n';
 import apiFetch from '@wordpress/api-fetch';
 import { useSelect } from '@wordpress/data';
 import {
-	JetpackFooter,
-	JetpackLogo,
+	AdminPage,
+	AdminSection,
+	AdminSectionHero,
+	Row,
+	Col,
 	getRedirectUrl,
 	PricingCard,
 } from '@automattic/jetpack-components';
@@ -80,40 +83,38 @@ const Admin = () => {
 			'jetpack-backup'
 		);
 		return (
-			<div className="jp-wrap jp-content">
-				<div className="jp-row jp-product-promote">
-					<div class="lg-col-span-6 md-col-span-6 sm-col-span-4">
-						<h1>{ __( 'Secure your site with a Backup subscription.', 'jetpack-backup' ) }</h1>
-						<p>
-							{ ' ' }
-							{ __(
-								'Get peace of mind knowing that all your work will be saved, and get back online quickly with one-click restores.',
-								'jetpack-backup'
-							) }
-						</p>
-						<ul>
-							<li>{ __( 'Automated real-time backups', 'jetpack-backup' ) }</li>
-							<li>{ __( 'Easy one-click restores', 'jetpack-backup' ) }</li>
-							<li>{ __( 'Complete list of all site changes', 'jetpack-backup' ) }</li>
-							<li>{ __( 'Global server infrastructure', 'jetpack-backup' ) }</li>
-							<li>{ __( 'Best-in-class support', 'jetpack-backup' ) }</li>
-						</ul>
-					</div>
-					<div class="lg-col-span-1 md-col-span-1 sm-col-span-0"></div>
-					<div class="lg-col-span-5 md-col-span-6 sm-col-span-4">
-						<PricingCard
-							ctaText={ __( 'Get Jetpack Backup', 'jetpack-backup' ) }
-							icon="data:image/svg+xml,%3Csvg width='32' height='32' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='m21.092 15.164.019-1.703v-.039c0-1.975-1.803-3.866-4.4-3.866-2.17 0-3.828 1.351-4.274 2.943l-.426 1.524-1.581-.065a2.92 2.92 0 0 0-.12-.002c-1.586 0-2.977 1.344-2.977 3.133 0 1.787 1.388 3.13 2.973 3.133H22.399c1.194 0 2.267-1.016 2.267-2.4 0-1.235-.865-2.19-1.897-2.368l-1.677-.29Zm-10.58-3.204a4.944 4.944 0 0 0-.201-.004c-2.75 0-4.978 2.298-4.978 5.133s2.229 5.133 4.978 5.133h12.088c2.357 0 4.267-1.97 4.267-4.4 0-2.18-1.538-3.99-3.556-4.339v-.06c0-3.24-2.865-5.867-6.4-5.867-2.983 0-5.49 1.871-6.199 4.404Z' fill='%23000'/%3E%3C/svg%3E"
-							infoText={ priceAfter === price ? basicInfoText : introductoryInfoText }
-							// eslint-disable-next-line react/jsx-no-bind
-							onCtaClick={ sendToCart }
-							priceAfter={ priceAfter }
-							priceBefore={ price }
-							title={ __( 'Jetpack Backup', 'jetpack-backup' ) }
-						/>
-					</div>
-				</div>
-			</div>
+			<Row>
+				<Col lg={ 6 } md={ 6 } sm={ 4 }>
+					<h1>{ __( 'Secure your site with a Backup subscription.', 'jetpack-backup' ) }</h1>
+					<p>
+						{ ' ' }
+						{ __(
+							'Get peace of mind knowing that all your work will be saved, and get back online quickly with one-click restores.',
+							'jetpack-backup'
+						) }
+					</p>
+					<ul className="jp-product-promote">
+						<li>{ __( 'Automated real-time backups', 'jetpack-backup' ) }</li>
+						<li>{ __( 'Easy one-click restores', 'jetpack-backup' ) }</li>
+						<li>{ __( 'Complete list of all site changes', 'jetpack-backup' ) }</li>
+						<li>{ __( 'Global server infrastructure', 'jetpack-backup' ) }</li>
+						<li>{ __( 'Best-in-class support', 'jetpack-backup' ) }</li>
+					</ul>
+				</Col>
+				<Col lg={ 1 } md={ 1 } sm={ 0 } />
+				<Col lg={ 5 } md={ 6 } sm={ 4 }>
+					<PricingCard
+						ctaText={ __( 'Get Jetpack Backup', 'jetpack-backup' ) }
+						icon="data:image/svg+xml,%3Csvg width='32' height='32' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill-rule='evenodd' clip-rule='evenodd' d='m21.092 15.164.019-1.703v-.039c0-1.975-1.803-3.866-4.4-3.866-2.17 0-3.828 1.351-4.274 2.943l-.426 1.524-1.581-.065a2.92 2.92 0 0 0-.12-.002c-1.586 0-2.977 1.344-2.977 3.133 0 1.787 1.388 3.13 2.973 3.133H22.399c1.194 0 2.267-1.016 2.267-2.4 0-1.235-.865-2.19-1.897-2.368l-1.677-.29Zm-10.58-3.204a4.944 4.944 0 0 0-.201-.004c-2.75 0-4.978 2.298-4.978 5.133s2.229 5.133 4.978 5.133h12.088c2.357 0 4.267-1.97 4.267-4.4 0-2.18-1.538-3.99-3.556-4.339v-.06c0-3.24-2.865-5.867-6.4-5.867-2.983 0-5.49 1.871-6.199 4.404Z' fill='%23000'/%3E%3C/svg%3E"
+						infoText={ priceAfter === price ? basicInfoText : introductoryInfoText }
+						// eslint-disable-next-line react/jsx-no-bind
+						onCtaClick={ sendToCart }
+						priceAfter={ priceAfter }
+						priceBefore={ price }
+						title={ __( 'Jetpack Backup', 'jetpack-backup' ) }
+					/>
+				</Col>
+			</Row>
 		);
 	};
 
@@ -128,11 +129,11 @@ const Admin = () => {
 			}
 
 			return (
-				<div className="jp-wrap">
-					<div className="jp-row">
-						<div class="lg-col-span-12 md-col-span-8 sm-col-span-4">{ renderConnectScreen() }</div>
-					</div>
-				</div>
+				<Row>
+					<Col lg={ 12 } md={ 8 } sm={ 4 }>
+						{ renderConnectScreen() }
+					</Col>
+				</Row>
 			);
 		}
 
@@ -151,54 +152,23 @@ const Admin = () => {
 
 		// Render an error state, this shouldn't occurr since we've passed userConnected checks
 		if ( capabilitiesError ) {
-			return <div>{ capabilitiesError }</div>;
+			return (
+				<Row>
+					<Col lg={ 12 } md={ 8 } sm={ 4 }>
+						{ capabilitiesError }
+					</Col>
+				</Row>
+			);
 		}
 
 		return renderNoBackupCapabilities();
 	};
 
-	const renderHeader = () => {
-		if ( showHeaderFooter ) {
-			return (
-				<div className="jp-wrap jp-content">
-					<div className="jp-row">
-						<div class="lg-col-span-12 md-col-span-8 sm-col-span-4">
-							<div className="jp-masthead">
-								<div className="jp-masthead__inside-container">
-									<div className="jp-masthead__logo-container">
-										<JetpackLogo className="jetpack-logo__masthead" />
-									</div>
-								</div>
-							</div>
-						</div>
-					</div>
-				</div>
-			);
-		}
-	};
-
-	const renderFooter = () => {
-		if ( showHeaderFooter ) {
-			return (
-				<div className="jp-wrap jp-content">
-					<div className="jp-row">
-						<div class="lg-col-span-12 md-col-span-8 sm-col-span-4">
-							<JetpackFooter
-								moduleName={ __( 'Jetpack Backup', 'jetpack-backup' ) }
-								a8cLogoHref="https://www.jetpack.com"
-							/>
-						</div>
-					</div>
-				</div>
-			);
-		}
-	};
-
 	// Renders additional segments under the jp-hero area condition on having a backup plan
 	const renderBackupSegments = () => {
 		return (
-			<div className="jp-row jp-content">
-				<div class="lg-col-span-6 md-col-span-4 sm-col-span-4">
+			<Row>
+				<Col lg={ 6 } md={ 4 }>
 					<h2>{ __( 'Where are my backups stored?', 'jetpack-backup' ) }</h2>
 					<p>
 						{ __(
@@ -220,9 +190,9 @@ const Admin = () => {
 							<span>{ __( 'Consider upgrading to real-time protection', 'jetpack-backup' ) }</span>
 						</a>
 					) }
-				</div>
-				<div class="lg-col-span-1 md-col-span-1 sm-col-span-0"></div>
-				<div class="lg-col-span-5 md-col-span-3 sm-col-span-4">
+				</Col>
+				<Col lg={ 1 } md={ 1 } sm={ 0 } />
+				<Col lg={ 5 } md={ 3 } sm={ 4 }>
 					<h2>{ __( "Your site's heartbeat", 'jetpack-backup' ) }</h2>
 					<p>
 						{ __(
@@ -243,28 +213,31 @@ const Admin = () => {
 					) }
 
 					{ renderConnectionStatusCard() }
-				</div>
-			</div>
+				</Col>
+			</Row>
 		);
 	};
 
 	const renderContent = () => {
 		return (
 			<div className="content">
-				<div className="jp-hero">{ renderLoadedState() }</div>
-				<div className="jp-section">
-					<div className="jp-wrap">{ isFullyConnected() && renderBackupSegments() }</div>
-				</div>
+				<AdminSectionHero>{ renderLoadedState() }</AdminSectionHero>
+				<AdminSection>{ isFullyConnected() && renderBackupSegments() }</AdminSection>
 			</div>
 		);
 	};
 
 	return (
-		<div id="jetpack-backup-admin-container">
-			{ renderHeader() }
-			{ renderContent() }
-			{ renderFooter() }
-		</div>
+		<AdminPage
+			withHeader={ showHeaderFooter }
+			withFooter={ showHeaderFooter }
+			moduleName={ __( 'Jetpack Backup', 'jetpack-backup' ) }
+			a8cLogoHref="https://www.jetpack.com"
+		>
+			<div id="jetpack-backup-admin-container" className="jp-content">
+				{ renderContent() }
+			</div>
+		</AdminPage>
 	);
 };